The override mechanism described in the spec has a pretty major limitation: in a cmp.h5 there is no way to override some movies but not others.
Additionally the current loadChemistry implementation does not honor
chemistry overrides in the bas/bax files---this could be fixed but
then we have to deal with the problem above somehow.

Override using the barcode triples in the h5 files.  The original
chemistry information can always be retrieved from the metadata.xml
file.